requirements.txt
pyromod
pyrogram>=2.0.106
tgcrypto
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
flask
